вся работа с избранным вынесена в отдельное приложение (в угоду DjCS)
