+++ data/fixtures.json
@@ -179,7 +179,7 @@
         }
     },
     {
-        "model": "goods.favorite",
+        "model": "favorites.favorite",
         "pk": 1,
         "fields": {
             "product": 5,
@@ -187,7 +187,7 @@
         }
     },
     {
-        "model": "goods.favorite",
+        "model": "favorites.favorite",
         "pk": 2,
         "fields": {
             "product": 1,
@@ -195,7 +195,7 @@
         }
     },
     {
-        "model": "goods.favorite",
+        "model": "favorites.favorite",
         "pk": 3,
         "fields": {
             "product": 2,
