+++ goods/admin.py
@@ -1,21 +1,7 @@
 from django.contrib import admin
 
-from goods.models import Category, Favorite, Product
+from goods.models import Category, Product
 
-class FavoriteTab(admin.TabularInline):
-    model = Favorite
-    fields = ("product", "product_description", "product_sell_price")
-    readonly_fields = ("product", "product_description","product_sell_price")
-    extra = 0
-
-    def product_description(self, obj):
-        return obj.product.description
-
-    def product_sell_price(self, obj):
-        return obj.product.sell_price
-    
-    product_description.short_description = "Описание продукта"
-    product_sell_price.short_description = "Цена"
 
 @admin.register(Category)
 class AdminCategory(admin.ModelAdmin):
@@ -40,4 +26,3 @@ class AdminProduct(admin.ModelAdmin):
         "is_published",
         "image",
     )
-
