+++ goods/models.py
@@ -65,28 +65,3 @@ class Product(BaseGoodModel):
         discount_amount = (self.price * self.discount) / 100
         return self.price - discount_amount
 
-
-class Favorite(models.Model):
-    product = models.ForeignKey(
-        to=Product,
-        on_delete=models.CASCADE,
-        related_name="favorites",
-        verbose_name="Продукт",
-    )
-    user = models.ForeignKey(
-        to=User,
-        on_delete=models.CASCADE,
-        related_name="favorites",
-        verbose_name="Пользователь",
-    )
-
-    class Meta:
-        db_table = "favorite"
-        verbose_name = "Избранное"
-        verbose_name_plural = "Избранные"
-        ordering = ("user__username",)
-        constraints = (
-            models.UniqueConstraint(
-                fields=("user", "product"), name="unique_favorite_product"
-            ),
-        )
