+++ goods/views.py
@@ -1,13 +1,11 @@
 from django.http import Http404
-from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models import Avg
-from django.shortcuts import get_object_or_404, redirect
-from django.urls import reverse
-from django.views import View
+from django.shortcuts import get_object_or_404
+
 from django.views.generic import ListView, DetailView
 
 from comments.forms import CommentForm
-from goods.models import Category, Favorite, Product
+from goods.models import Category, Product
 
 
 class CatalogView(ListView):
@@ -48,7 +46,7 @@ class ProductView(DetailView):
         context = super().get_context_data(**kwargs)
         context["title"] = f"CoffeeShop - {self.object.name}"
         context["comment_count"] = self.object.comments.count()
-        context["average_rating"] = self.average_rating() #self.object.comments.aggregate(Avg('rating'))['rating__avg']
+        context["average_rating"] = self.average_rating()
         context["is_favorite"] = self.is_favorite()
         context["comment_form"] = CommentForm()  
 
@@ -64,32 +62,3 @@ class ProductView(DetailView):
         if rating:
             return round(rating, 1)
         return "Пока нет оценок"
-
-# Представления работы с избранным
-class AddFavorite(LoginRequiredMixin, View):
-    def post(self, request, product_slug):
-        if request.user.favorites.filter(product__slug = product_slug).exists():
-            raise ValueError("Товар уже в избранном")
-        
-        Favorite.objects.create(
-            user = request.user,
-            product = Product.objects.get(slug=product_slug)
-        )
-
-        return redirect(
-            "catalog:product", product_slug = product_slug
-            )
-
-class DeleteFavorite(LoginRequiredMixin, View):
-    def post(self, request, product_slug):
-        favorite_obj = request.user.favorites.filter(product__slug = product_slug).first()
-        print(favorite_obj)
-
-        if not favorite_obj:
-            raise ValueError("Товара нет в избранном")
-        favorite_obj.delete()
-
-        return redirect(
-            "catalog:product",
-            product_slug = product_slug
-            )
