+++ readme_docs/done.md
@@ -24,5 +24,6 @@
   + в карточке товара ведётся подсчёт отзывов к товару
 
 6. Механизм избранных товаров.
+  + вынести избранное в отдельное приложение
   + пользователь может добавить товары в избранное, убрать из избранного
   + в профиле отображаются товары добавленные в избранное
