+++ readme_docs/task.md
@@ -3,7 +3,6 @@
   - отсортировать товары по популярности, по рейтингу, по цене >, по цене от <
   - отфильтровать товары по бренду, цене от-до
 
-
 2. Функционал пользователя:
   - Изменить пароль
   - Восстановить пароль
@@ -59,6 +58,16 @@
   - добавить поле поиска в шаблоны
   - добавить выделения найденного текста
 
-13. CI\CD и докеризация
+13. Автоматическое тестирование pytest:
+  - написать тесты для моделей продуктов - добавление \ отображение 
+  - написать тесты для корзин -   добавление и удаление (авториз\аноним)
+  - написать тесты для заказов - оформление (авториз\аноним)
+  - написать тесты для комментариев - добавление \ удаление \ ? изменение (авториз\аноним)
+
+14. CI\CD и докеризация
   - написать контейнеры
   - создать пайплайн
+
+*15. Форма обратной связи:
+  - сделать иконку (задать вопрос\  напиши нашему менеджеру), которая перекинет в рабочий чат в ТГ или ВА
+  \ откроет форму чата
