--- templates/carts/included_cart.html
@@ -1,59 +1,0 @@
-{% load static %}
-
-<div class="card mb-3 text-bg-light shadow-lg">
-    {% for cart in carts %}
-        <div class="card-header">
-            <h5 class="card-title">{{ cart.product.name }}</h5>
-        </div>
-        <ul class="list-group list-group-flush">
-            <li class="list-group-item">
-                <div class="row text-center">
-                    <div class="col p-0">
-                        <div class="input-group">
-                            <span class="input-group-btn">
-                                <button type="button" class="btn btn-dark btn-sm decrement"
-                                data-cart-id="{{ cart.id }}"
-                                data-cart-change-url="{% url "cart:cart_change" %}">
-                                {% csrf_token %}
-                                -</button>
-                            </span>
-                            <input type="text" class="form-control number" value="{{ cart.amount }}"
-                                readonly>
-                            <span class="input-group-btn">
-                                <button type="button" class="btn btn-dark btn-sm increment"
-                                data-cart-id="{{ cart.id }}"
-                                data-cart-change-url="{% url "cart:cart_change" %}">
-                                {% csrf_token %}
-                                +</button>
-                            </span>
-                        </div>
-                    </div>
-                    <div class="col p-0">
-                        <p>x {{ cart.product.sell_price }} = </p>
-                    </div>
-                    <div class="col p-0"><strong> {{cart.products_price}} руб. </strong></div>
-                    <div class="col p-0">
-                        <a href="{% url "cart:cart_remove" %}" class="remove-from-cart" data-cart-id="{{ cart.id }}">
-                        {% csrf_token %}
-                            <img class="mx-1" src="{% static "deps/icons/trash3-fill.svg" %}"
-                                alt="Catalog Icon" width="16" height="16">
-                        </a>
-                    </div>
-                </div>
-            </li>
-        </ul>
-    {% endfor %}
-</div>
-<div class="card mb-3 shadow-lg">
-    <div class="card-footer">
-        <p class="float-left">Итого <strong>{{ carts.total_amount }}</strong> товар(а) на сумму</p>
-        <h4 class="float-left"><strong>{{ carts.total_price }} руб. </strong></h4>
-    </div>
-</div>
-
-
-{% if carts and not order%}
-    <a class="btn btn-dark" href="{% url "orders:create_order" %}">
-        Оформить заказ
-    </a>
-{% endif %}
